Refactor and add icons
